+++ config.yaml
@@ -1,3 +1,8 @@
 [General]
 auroraThresh: 30
 announcements: 1281013297847603281
+quotes: 1280656734196596858
+debug: 1280991837913481278
+food: 1358906415426830387
+organize: 772510418920144936
+actionlog: 1305223850153480245
